requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
openai>=1.0.0
sentence-transformers>=2.2.2 